# docs/chainhook-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "lumaLang",
    "version": "1.8.0"
  },
  "paths": {
...
    "/v1/lumalanghooks": {
      "get": {
        "tags": [
          "Managing Predicates"
        ],
        "operationId": "handle_get_predicates",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Managing Predicates"
        ],
        "operationId": "handle_create_predicate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LumalanghookSpecificationNetworkMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/lumalanghooks/{predicate_uuid}": {
      "get": {
        "tags": [
          "Managing Predicates"
        ],
        "operationId": "handle_get_predicate",
        "parameters": [
          {
            "name": "predicate_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/lumalanghooks/bitcoin/{predicate_uuid}": {
      "delete": {
        "tags": [
          "Managing Predicates"
        ],
        "operationId": "handle_delete_bitcoin_predicate",
        "parameters": [
          {
            "name": "predicate_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/lumalanghooks/stacks/{predicate_uuid}": {
      "delete": {
        "tags": [
          "Managing Predicates"
        ],
        "operationId": "handle_delete_stacks_predicate",
        "parameters": [
          {
            "name": "predicate_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LumalanghookSpecificationNetworkMap": {
        "oneOf": [
          {
            "description": "Maps some [BitcoinLumalanghookSpecification] to a corresponding [BitcoinNetwork]. This allows maintaining one serialized predicate file for a given predicate on each network.\n\n### Examples Given some file `predicate.json`: ```json { \"uuid\": \"my-id\", \"name\": \"My Predicate\", \"chain\": \"bitcoin\", \"version\": 1, \"networks\": { \"regtest\": { // ... }, \"testnet\": { // ... }, \"mainnet\": { // ... } } } ``` You can deserialize the file to this type and create a [BitcoinLumalanghookInstance] for the desired network: ``` use lumalanghook_sdk::lumalanghooks::bitcoin::BitcoinLumalanghookSpecificationNetworkMap; use lumalanghook_sdk::lumalanghooks::bitcoin::BitcoinLumalanghookInstance; use lumalanghook_types::BitcoinNetwork;\n\nfn get_predicate(network: &BitcoinNetwork) -> Result<BitcoinLumalanghookInstance, String> { let json_predicate = std::fs::read_to_string(\"./predicate.json\").expect(\"Unable to read file\"); let hook_map: BitcoinLumalanghookSpecificationNetworkMap = serde_json::from_str(&json_predicate).expect(\"Unable to parse LumaLang map\"); hook_map.into_specification_for_network(network) }\n\n```",
            "type": "object",
            "required": [
              "chain",
              "name",
              "networks",
              "uuid",
              "version"
            ],
            "properties": {
              "chain": {
                "type": "string",
                "enum": [
                  "bitcoin"
                ]
              },
              "uuid": {
                "type": "string"
              },
              "owner_uuid": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              },
              "version": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              },
              "networks": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": [
                      "regtest"
                    ],
                    "properties": {
                      "regtest": {
                        "$ref": "#/components/schemas/BitcoinLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "testnet"
                    ],
                    "properties": {
                      "testnet": {
                        "$ref": "#/components/schemas/BitcoinLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "signet"
                    ],
                    "properties": {
                      "signet": {
                        "$ref": "#/components/schemas/BitcoinLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "mainnet"
                    ],
                    "properties": {
                      "mainnet": {
                        "$ref": "#/components/schemas/BitcoinLumalanghookSpecification"
                      }
                    }
                  }
                ]
              }
            }
          },
          {
            "description": "Maps some [StacksLumalanghookSpecification] to a corresponding [StacksNetwork]. This allows maintaining one serialized predicate file for a given predicate on each network.\n\n### Examples Given some file `predicate.json`: ```json { \"uuid\": \"my-id\", \"name\": \"My Predicate\", \"chain\": \"stacks\", \"version\": 1, \"networks\": { \"devnet\": { // ... }, \"testnet\": { // ... }, \"mainnet\": { // ... } } } ``` You can deserialize the file to this type and create a [StacksLumalanghookInstance] for the desired network: ``` use lumalanghook_sdk::lumalanghooks::stacks::StacksLumalanghookSpecificationNetworkMap; use lumalanghook_sdk::lumalanghooks::stacks::StacksLumalanghookInstance; use lumalanghook_types::StacksNetwork;\n\nfn get_predicate(network: &StacksNetwork) -> Result<StacksLumalanghookInstance, String> { let json_predicate = std::fs::read_to_string(\"./predicate.json\").expect(\"Unable to read file\"); let hook_map: StacksLumalanghookSpecificationNetworkMap = serde_json::from_str(&json_predicate).expect(\"Unable to parse LumaLang map\"); hook_map.into_specification_for_network(network) }\n\n```",
            "type": "object",
            "required": [
              "chain",
              "name",
              "networks",
              "uuid",
              "version"
            ],
            "properties": {
              "chain": {
                "type": "string",
                "enum": [
                  "stacks"
                ]
              },
              "uuid": {
                "type": "string"
              },
              "owner_uuid": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              },
              "version": {
                "type": "integer",
                "format": "uint32",
                "minimum": 0.0
              },
              "networks": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": [
                      "simnet"
                    ],
                    "properties": {
                      "simnet": {
                        "$ref": "#/components/schemas/StacksLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "devnet"
                    ],
                    "properties": {
                      "devnet": {
                        "$ref": "#/components/schemas/StacksLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "testnet"
                    ],
                    "properties": {
                      "testnet": {
                        "$ref": "#/components/schemas/StacksLumalanghookSpecification"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "mainnet"
                    ],
                    "properties": {
                      "mainnet": {
                        "$ref": "#/components/schemas/StacksLumalanghookSpecification"
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
...
      "BitcoinLumalanghookSpecification": {
        "type": "object",
        "required": [
          "if_this",
          "then_that"
        ],
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0.0
            },
            "nullable": true
          },
          "start_block": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "end_block": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "expire_after_occurrence": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "include_proof": {
            "type": "boolean",
            "nullable": true
          },
          "include_inputs": {
            "type": "boolean",
            "nullable": true
          },
          "include_outputs": {
            "type": "boolean",
            "nullable": true
          },
          "include_witness": {
            "type": "boolean",
            "nullable": true
          },
          "if_this": {
            "$ref": "#/components/schemas/BitcoinPredicateType"
          },
          "then_that": {
            "$ref": "#/components/schemas/HookAction"
          }
        }
      },
      "BitcoinPredicateType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "block"
                ]
              }
            }
          },
          {
            "type": "object",
            "oneOf": [
              {
                "type": "object",
                "required": [
                  "equals"
                ],
                "properties": {
                  "equals": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            ],
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "txid"
                ]
              }
            }
          },
          {
            "type": "object",
            "oneOf": [
              {
                "type": "object",
                "required": [
                  "txid"
                ],
                "properties": {
                  "txid": {
                    "$ref": "#/components/schemas/TxinPredicate"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "witness_script"
                ],
                "properties": {
                  "witness_script": {
                    "$ref": "#/components/schemas/MatchingRule"
                  }
                },
                "additionalProperties": false
              }
            ],
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "inputs"
                ]
              }
            }
          },
          {
            "type": "object",
            "oneOf": [
              {
                "type": "object",
                "required": [
                  "op_return"
                ],
                "properties": {
                  "op_return": {
                    "$ref": "#/components/schemas/MatchingRule"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "p2pkh"
                ],
                "properties": {
                  "p2pkh": {
                    "$ref": "#/components/schemas/ExactMatchingRule"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "p2sh"
                ],
                "properties": {
                  "p2sh": {
                    "$ref": "#/components/schemas/ExactMatchingRule"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "p2wpkh"
                ],
                "properties": {
                  "p2wpkh": {
                    "$ref": "#/components/schemas/ExactMatchingRule"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "p2wsh"
                ],
                "properties": {
                  "p2wsh": {
                    "$ref": "#/components/schemas/ExactMatchingRule"
                  }
                },
                "additionalProperties": false
              },
              {
                "type": "object",
                "required": [
                  "descriptor"
                ],
                "properties": {
                  "descriptor": {
                    "$ref": "#/components/schemas/DescriptorMatchingRule"
                  }
                },
                "additionalProperties": false
              }
            ],
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "outputs"
                ]
              }
            }
          },
          {
            "type": "object",
            "oneOf": [
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "stacker_rewarded"
                    ]
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "block_committed"
                    ]
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "leader_registered"
                    ]
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "stx_transferred"
                    ]
                  }
                }
              },
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "stx_locked"
                    ]
                  }
                }
              }
            ],
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "stacks_protocol"
                ]
              }
            }
          },
          {
            "type": "object",
            "oneOf": [
              {
                "type": "object",
                "required": [
                  "operation"
                ],
                "properties": {
                  "operation": {
                    "type": "string",
                    "enum": [
                      "inscription_feed"
                    ]
                  },
                  "meta_protocols": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/OrdinalsMetaProtocol"
                    },
                    "uniqueItems": true,
                    "nullable": true
                  }
                }
              }
            ],
            "required": [
              "scope"
            ],
            "properties": {
              "scope": {
                "type": "string",
                "enum": [
                  "ordinals_protocol"
                ]
              }
            }
          }
        ]
      },
      "TxinPredicate": {
        "type": "object",
        "required": [
          "txid",
          "vout"
        ],
        "properties": {
          "txid": {
            "type": "string"
          },
          "vout": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          }
        }
      },
      "MatchingRule": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "equals"
            ],
            "properties": {
              "equals": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": [
              "starts_with"
            ],
            "properties": {
              "starts_with": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": [
              "ends_with"
            ],
            "properties": {
              "ends_with": {
                "type": "string"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "ExactMatchingRule": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "equals"
            ],
            "properties": {
              "equals": {
                "type": "string"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "OrdinalsMetaProtocol": {
        "type": "string",
        "enum": [
          "all",
          "brc-20"
        ]
      },
      "HookAction": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "noop"
            ]
          },
          {
            "type": "object",
            "required": [
              "http_post"
            ],
            "properties": {
              "http_post": {
                "$ref": "#/components/schemas/HttpHook"
              }
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": [
              "file_append"
            ],
            "properties": {
              "file_append": {
                "$ref": "#/components/schemas/FileHook"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "FileHook": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
...
      "StacksLumalanghookSpecification": {
        "type": "object",
        "required": [
          "if_this",
          "then_that"
        ],
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0.0
            },
            "nullable": true
          },
          "start_block": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "end_block": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "expire_after_occurrence": {
            "type": "integer",
            "format": "uint64",
            "minimum": 0.0,
            "nullable": true
          },
          "capture_all_events": {
            "type": "boolean",
            "nullable": true
          },
          "decode_clarity_values": {
            "type": "boolean",
            "nullable": true
          },
          "include_contract_abi": {
            "type": "boolean",
            "nullable": true
          },
          "if_this": {
            "$ref": "#/components/schemas/StacksPredicate"
          },
          "then_that": {
            "$ref": "#/components/schemas/HookAction"
          }
        }
      },
...
    }
  }
}